requirements.txt
django~=3.1
djangorestframework~=3.12
gunicorn~=20.0
requests~=2.25